pyproject.toml
[project]
name = "data_management_agent"
version = "0.1.0"
description = "Data Management Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.15.0",
    "mcp>=1.18.0",
    "toolbox-core>=0.5.2"
]
